Patch 12.2.3327
I think. Confirmed I have the latest available sgas and ran the commands
in the readme. No manual cleanup on any of the data, just the automated results.

Wanted to see what the diff looked like and maybe use this as a start
for more discussion as I onboard with contributing.

Ran `yarn build` to update the civ and unified json.